Corrijo errores

+++ src/main/java/com/omega/Proyecto/omega/Model/Employee.java
@@ -13,7 +13,6 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Employee extends Person{
-
     private String position;
-    private Double salary;
+    private Long salary;
 }
